refactor(desktop): reduce integration to stdio frontend
